# Improve performance of `dnp.nan_to_num`