… dpnp/backend/extensions/ufunc/CMakeLists.txt
… dpnp/backend/extensions/ufunc/elementwise_functions/nan_to_num.hpp
… dpnp/dpnp_iface_mathematical.py
… dpnp/tests/test_mathematical.py