@ndgrigorian 2025-02-05
Also, for posterity, the tests all pass on CUDA